Refactor some
