+++ src/config.ts
@@ -1,0 +1,1 @@
+export const TRANSITION_TIME = 30000;
